+++ handler_success.go
@@ -25,6 +25,7 @@ const (
 	VersionKeyContext = 3
 	OrgSessionContext = 4
 	ContextData       = 5
+	RetainHost        = 6
 )
 
 var SessionCache *cache.Cache = cache.New(10*time.Second, 5*time.Second)